FEATURE: CL-8490 Migrate jQuery AJAX to fetch api (#116)
Authors: Adam Klouboucnik, Vojtech Jasny, Vaclav Bohac

+++ src/user.js
@@ -1,6 +1,5 @@
 // Copyright (C) 2007-2014, GoodData(R) Corporation. All rights reserved.
-import $ from 'jquery';
-import { ajax, get, post, put } from './xhr';
+import { ajax, get, post, put, parseJSON } from './xhr';
 
 /**
  * @module user
@@ -10,13 +9,27 @@ import { ajax, get, post, put } from './xhr';
 /**
  * Find out whether a user is logged in
  *
- * Returns a promise which either:
- * **resolves** - which means user is logged in or
- * **rejects** - meaning is not logged in
+ * @return {Promise} resolves with true if user logged in, false otherwise
  * @method isLoggedIn
  */
 export function isLoggedIn() {
-    return $.getJSON('/gdc/account/token');
+    return new Promise((resolve, reject) => {
+        // cannot use get here directly - we need to access to response
+        // not to responses JSON get returns
+        ajax('/gdc/account/token', { method: 'GET' }).then(r => {
+            if (r.ok) {
+                resolve(true);
+            }
+
+            resolve(false);
+        }, err => {
+            if (err.response.status === 401) {
+                resolve(false);
+            } else {
+                reject(err);
+            }
+        });
+    });
 }
 
 
@@ -31,7 +44,7 @@ export function isLoggedIn() {
  */
 export function login(username, password) {
     return post('/gdc/account/login', {
-        data: JSON.stringify({
+        body: JSON.stringify({
             postUserLogin: {
                 login: username,
                 password: password,
@@ -40,7 +53,7 @@ export function login(username, password) {
                 verifyCaptcha: ''
             }
         })
-    });
+    }).then(parseJSON);
 }
 
 /**
@@ -48,23 +61,20 @@ export function login(username, password) {
  * @method logout
  */
 export function logout() {
-    /* eslint new-cap: 0 */
-    const d = $.Deferred();
+    return isLoggedIn().then((loggedIn) => {
+        if (loggedIn) {
+            return get('/gdc/app/account/bootstrap').then((result) => {
+                const userUri = result.bootstrapResource.accountSetting.links.self;
+                const userId = userUri.match(/([^\/]+)\/?$/)[1];
 
-    isLoggedIn().then(function resolve() {
-        return get('/gdc/app/account/bootstrap').then(function resolveGet(result) {
-            const userUri = result.bootstrapResource.accountSetting.links.self;
-            const userId = userUri.match(/([^\/]+)\/?$/)[1];
+                return ajax('/gdc/account/login/' + userId, {
+                    method: 'delete'
+                });
+            });
+        }
 
-            return userId;
-        }, d.reject);
-    }, d.resolve).then(function resolveAll(userId) {
-        return ajax('/gdc/account/login/' + userId, {
-            method: 'delete'
-        });
-    }).then(d.resolve, d.reject);
-
-    return d.promise();
+        return Promise.resolve();
+    });
 }
 
 /**
@@ -74,15 +84,9 @@ export function logout() {
  * @param {Object} profileSetting
 */
 export function updateProfileSettings(profileId, profileSetting) {
-    /*eslint-disable new-cap*/
-    const d = $.Deferred();
-    /*eslint-enable new-cap*/
-
-    put('/gdc/account/profile/' + profileId + '/settings', {
+    return put('/gdc/account/profile/' + profileId + '/settings', {
         data: profileSetting
-    }).then(d.resolve, d.reject);
-
-    return d.promise();
+    });
 }
 
 /**
@@ -90,22 +94,18 @@ export function updateProfileSettings(profileId, profileSetting) {
  * @method getAccountInfo
  */
 export function getAccountInfo() {
-    /* eslint new-cap: 0 */
-    const d = $.Deferred();
+    return get('/gdc/app/account/bootstrap')
+        .then(function resolveBootstrap(result) {
+            const br = result.bootstrapResource;
+            const accountInfo = {
+                login: br.accountSetting.login,
+                loginMD5: br.current.loginMD5,
+                firstName: br.accountSetting.firstName,
+                lastName: br.accountSetting.lastName,
+                organizationName: br.settings.organizationName,
+                profileUri: br.accountSetting.links.self
+            };
 
-    get('/gdc/app/account/bootstrap').then(function resolveBootstrap(result) {
-        const br = result.bootstrapResource;
-        const accountInfo = {
-            login: br.accountSetting.login,
-            loginMD5: br.current.loginMD5,
-            firstName: br.accountSetting.firstName,
-            lastName: br.accountSetting.lastName,
-            organizationName: br.settings.organizationName,
-            profileUri: br.accountSetting.links.self
-        };
-
-        d.resolve(accountInfo);
-    }, d.reject);
-
-    return d.promise();
+            return accountInfo;
+        });
 }
